requirements-async.txt
# Core async framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Database (async)
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0  # For SQLite (dev)
asyncpg>=0.29.0    # For PostgreSQL (production)
alembic>=1.12.0    # Migrations

# Redis (async)
redis[hiredis]>=5.0.0

# HTTP client (async)
httpx>=0.25.0

# Twilio
twilio>=8.10.0

# AI & ML
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
Pillow>=10.1.0
python-magic>=0.4.27

# Vector Store
chromadb>=0.4.0

# Background Tasks
celery[redis]>=5.3.0

# Audio (Voice messages)
openai>=1.3.0  # For Whisper transcription
google-cloud-texttospeech>=2.14.0  # Optional

# Citations & References
bibtexparser>=1.4.0

# Utilities
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0  # For fallback

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0  # For testing

# Rate limiting
slowapi>=0.1.9

# Monitoring (optional)
sentry-sdk[fastapi]>=1.38.0

# Production server
gunicorn>=21.2.0
